Move rhythm difficulty calculation into a separate `Skill`
